+++ lib/active_observers.dart
@@ -1,8 +1,8 @@
 library active_observers;
 
-export 'src/active_observers.dart'
-    show ActiveObservers, StateLifecyclePhase;
+export 'src/active_observers.dart' show ActiveObservers, StateLifecyclePhase;
 export 'src/observe_effect.dart' show observeEffect;
 export 'src/observe_lifecycle.dart' show observeLifecycle;
 export 'src/observe_listenable.dart' show observeListenable;
+export 'src/observe_paint.dart' show observePaint;
 export 'src/observe_stream.dart' show observeStream;