+++ lib/src/observe_lifecycle.dart
@@ -1,5 +1,7 @@
 import 'active_observers.dart';
 
+export 'active_observers.dart' show StateLifecyclePhase;
+
 /// During every lifecycle of [State], [run] will get called and receive a
 /// [StateLifecyclePhase] representing current lifecycle.
 void observeLifecycle(void Function(StateLifecyclePhase) run) {